requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
Django==5.1.6
django-extensions==3.2.3
django-redis==5.4.0
django-solo==2.1.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
executing==2.2.0
idna==3.10
ipython
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-decouple==3.8
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.3.0
wcwidth==0.2.13
